requirements.txt
fastapi
uvicorn
twilio
openai
python-dotenv
requests
python-multipart
openai-whisper
torch